apps/server/requirements.txt
fastapi[all]==0.116.1
pydantic==2.11.7
pydantic-settings==2.6.1
bcrypt==4.1.2
pymongo==4.14.1
PyJWT==2.9.0
python-multipart>=0.0.18
pytest==8.4.1
pytest-asyncio==1.1.0
uvicorn[standard]==0.32.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1